requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
build==1.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==4.1.8
et-xmlfile==1.1.0
googlemaps==4.10.0
idna==3.7
importlib_metadata==8.0.0
mysqlclient==2.1.1
openpyxl==3.1.5
packaging==24.1
pip-tools==7.4.1
polyline==2.0.2
PyMySQL==1.1.1
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.2.2
zipp==3.19.2
